Refactor old .agents to subgraphs
