--- packages/@n8n/ai-workflow-builder.ee/src/agents/discovery.agent.ts
@@ -1,128 +1,0 @@
-import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
-import { ChatPromptTemplate } from '@langchain/core/prompts';
-import type { INodeTypeDescription } from 'n8n-workflow';
-
-import type { BuilderTool } from '@/utils/stream-processor';
-
-import { createNodeDetailsTool } from '../tools/node-details.tool';
-import { createNodeSearchTool } from '../tools/node-search.tool';
-
-/**
- * Discovery Agent Prompt
- *
- * Focused on finding and analyzing nodes needed for the workflow.
- * Does NOT build or configure - only discovers and recommends.
- */
-const discoveryAgentPrompt = `You are a Discovery Agent specialized in finding n8n workflow nodes.
-
-Your ONLY job is to:
-1. Understand what the user wants to build
-2. Search for ALL relevant NODE TYPES (not actual data/URLs)
-3. Get detailed information about those node types
-4. Report back ALL nodes you found with their capabilities
-
-WHAT YOU SEARCH FOR:
-- n8n NODE TYPES like "RSS Read", "OpenAI", "Gmail", "Schedule Trigger"
-- NOT actual RSS feed URLs, API endpoints, or data sources
-- Think: "What n8n integration nodes exist?" not "What websites have RSS feeds?"
-
-SEARCH STRATEGY:
-- Search broadly first (e.g., "OpenAI" not just "ChatGPT")
-- Search for ALL node types needed for the workflow
-- Call get_node_details for nodes that seem relevant
-- Search in PARALLEL when looking for multiple node types
-- Consider alternative nodes (e.g., both "HTTP Request" and service-specific nodes)
-
-CRITICAL RULES:
-- Call search_nodes to find node types by name
-- Call get_node_details to understand node inputs/outputs/parameters
-- Execute tools SILENTLY - no commentary before or between tool calls
-- Be efficient - you have a call limit
-- Find ALL matching nodes, not just the first one
-
-DO NOT:
-- Output text before calling tools
-- Add commentary between tool calls
-- Try to build the workflow (that's the Builder Agent's job)
-- Configure parameters (that's the Configurator Agent's job)
-- Make assumptions about which nodes to use - always search first
-- Stop after finding just one node - find all relevant options
-
-RESPONSE FORMAT:
-After ALL tools have completed, provide a comprehensive summary:
-
-**Nodes Found:**
-- [Node Display Name] ([Node Type]): [Key capability and why it's relevant]
-- [Node Display Name] ([Node Type]): [Key capability and why it's relevant]
-...
-
-**Additional Context:**
-- Any important details about node compatibility, credentials, or limitations
-- Recommendations for the builder (e.g., "Connect HTTP Request output to Code node")
-- Data format or API considerations
-
-Example:
-**Nodes Found:**
-- Schedule Trigger (scheduleTrigger): Triggers workflow on daily schedule
-- OpenWeatherMap (openWeatherMap): Fetches weather data by location
-- DALL-E (openAiDallE): Generates images from text prompts
-- Gmail (gmail): Sends emails with attachments
-
-**Additional Context:**
-- OpenWeatherMap requires API credentials
-- DALL-E node outputs image URLs, not binary data
-- Gmail can attach images via URL
-
-CRITICAL:
-- Only respond AFTER all search_nodes and get_node_details tools have executed
-- Include ALL nodes you found, not just a selection
-- Provide reasoning for WHY each node is relevant`;
-
-const systemPrompt = ChatPromptTemplate.fromMessages([
-	['system', discoveryAgentPrompt],
-	['placeholder', '{messages}'],
-]);
-
-export interface DiscoveryAgentConfig {
-	llm: BaseChatModel;
-	parsedNodeTypes: INodeTypeDescription[];
-}
-
-/**
- * Discovery Agent
- *
- * Responsible for finding and analyzing nodes from the n8n catalog.
- * Works with search_nodes and get_node_details tools.
- */
-export class DiscoveryAgent {
-	private llm: BaseChatModel;
-	private tools: BuilderTool[];
-
-	constructor(config: DiscoveryAgentConfig) {
-		this.llm = config.llm;
-
-		// Only bind tools relevant to discovery
-		this.tools = [
-			createNodeSearchTool(config.parsedNodeTypes),
-			createNodeDetailsTool(config.parsedNodeTypes),
-		];
-	}
-
-	/**
-	 * Get the tools for this agent
-	 */
-	getTools(): BuilderTool[] {
-		return this.tools;
-	}
-
-	/**
-	 * Get the agent's LLM with tools bound
-	 */
-	getAgent() {
-		const toolsForBinding = this.tools.map((bt) => bt.tool);
-		if (typeof this.llm.bindTools !== 'function') {
-			throw new Error('LLM does not support tools');
-		}
-		return systemPrompt.pipe(this.llm.bindTools(toolsForBinding));
-	}
-}
