+++ packages/@n8n/ai-workflow-builder.ee/src/agents/index.ts
@@ -1,6 +1,3 @@
-export { DiscoveryAgent, type DiscoveryAgentConfig } from './discovery.agent';
-export { BuilderAgent, type BuilderAgentConfig } from './builder.agent';
-export { ConfiguratorAgent, type ConfiguratorAgentConfig } from './configurator.agent';
 export { ResponderAgent, type ResponderAgentConfig } from './responder.agent';
 export {
 	SupervisorAgent,
