+++ packages/@n8n/ai-workflow-builder.ee/src/utils/stream-processor.ts
@@ -54,6 +54,9 @@ const NODES_TO_SKIP = [
 	'cleanup_dangling_tool_calls', // Internal cleanup
 	'create_workflow_name', // Internal workflow naming
 	'auto_compact_messages', // Internal state management
+	'configurator_subgraph',
+	'discovery_subgraph',
+	'builder_subgraph',
 ];
 
 /**
@@ -188,42 +191,11 @@ export function processStreamChunk(streamMode: string, chunk: unknown): StreamOu
 export async function* createStreamProcessor(
 	stream: AsyncGenerator<any, void, unknown>,
 ): AsyncGenerator<StreamOutput> {
-	let eventCount = 0;
 	for await (const event of stream) {
-		eventCount++;
-
-		console.log(`[Stream] Event #${eventCount}`, {
-			isArray: Array.isArray(event),
-			length: Array.isArray(event) ? event.length : 'N/A',
-			firstElemIsArray: Array.isArray(event) && Array.isArray(event[0]),
-			firstElemType: Array.isArray(event) ? typeof event[0] : 'N/A',
-			elements: Array.isArray(event)
-				? event.map((e, i) => ({
-						index: i,
-						type: typeof e,
-						isArray: Array.isArray(e),
-						value: Array.isArray(e) ? `array[${e.length}]` : String(e).substring(0, 50),
-					}))
-				: 'N/A',
-		});
-
 		// Subgraph events: flat 3-element array [namespace, streamMode, data]
 		if (Array.isArray(event) && event.length === 3 && Array.isArray(event[0])) {
-			const [namespace, streamMode, data] = event;
-			const namespaceStr = namespace.join('/') || 'parent';
-
-			console.log('[Stream] Processing subgraph event', {
-				namespace: namespaceStr,
-				streamMode,
-				dataKeys: typeof data === 'object' && data ? Object.keys(data) : 'N/A',
-			});
-
+			const [_namespace, streamMode, data] = event;
 			const output = processStreamChunk(streamMode, data);
-
-			console.log('[Stream] Subgraph output', {
-				hasOutput: !!output,
-				outputKeys: output ? Object.keys(output) : 'none',
-			});
 
 			if (output) {
 				yield output;
@@ -232,20 +204,9 @@ export async function* createStreamProcessor(
 			// Regular parent event: [streamMode, chunk]
 			const [streamMode, chunk] = event;
 
-			console.log('[Stream] Detected parent event', {
-				streamMode,
-				streamModeType: typeof streamMode,
-				streamModeLength: typeof streamMode === 'string' ? streamMode.length : 'N/A',
-			});
-
 			// Make sure it's a valid stream mode string
 			if (typeof streamMode === 'string' && streamMode.length > 1) {
 				const output = processStreamChunk(streamMode, chunk);
-
-				console.log('[Stream] Parent output', {
-					hasOutput: !!output,
-					outputKeys: output ? Object.keys(output) : 'none',
-				});
 
 				if (output) {
 					yield output;
@@ -253,7 +214,6 @@ export async function* createStreamProcessor(
 			}
 		}
 	}
-	console.log(`[Stream] Finished processing ${eventCount} events`);
 }
 
 /**
